refactor(amazon): de-angularize KeyPairsReader
