+++ app/scripts/modules/amazon/src/keyPairs/index.ts
@@ -1,1 +1,1 @@
-export * from './keyPairs.read.service';
+export * from './KeyPairsReader';
