+++ app/scripts/modules/amazon/src/serverGroup/configure/serverGroupConfiguration.service.spec.ts
@@ -10,7 +10,7 @@ import {
   SubnetReader,
 } from '@spinnaker/core';
 
-import { KeyPairsReader } from 'amazon/keyPairs/keyPairs.read.service';
+import { KeyPairsReader } from 'amazon/keyPairs';
 import {
   AWS_SERVER_GROUP_CONFIGURATION_SERVICE,
   AwsServerGroupConfigurationService,
@@ -22,7 +22,6 @@ describe('Service: awsServerGroupConfiguration', function() {
     securityGroupReader: SecurityGroupReader,
     awsInstanceTypeService: any,
     cacheInitializer: CacheInitializerService,
-    keyPairsReader: KeyPairsReader,
     loadBalancerReader: LoadBalancerReader,
     applicationModelBuilder: ApplicationModelBuilder,
     $scope: IScope;
@@ -36,7 +35,6 @@ describe('Service: awsServerGroupConfiguration', function() {
       _securityGroupReader_: SecurityGroupReader,
       _awsInstanceTypeService_: any,
       _cacheInitializer_: CacheInitializerService,
-      _keyPairsReader_: KeyPairsReader,
       _loadBalancerReader_: LoadBalancerReader,
       _applicationModelBuilder_: ApplicationModelBuilder,
       $rootScope: IRootScopeService,
@@ -46,7 +44,6 @@ describe('Service: awsServerGroupConfiguration', function() {
       securityGroupReader = _securityGroupReader_;
       awsInstanceTypeService = _awsInstanceTypeService_;
       cacheInitializer = _cacheInitializer_;
-      keyPairsReader = _keyPairsReader_;
       loadBalancerReader = _loadBalancerReader_;
       applicationModelBuilder = _applicationModelBuilder_;
       $scope = $rootScope.$new();
@@ -97,7 +94,7 @@ describe('Service: awsServerGroupConfiguration', function() {
       );
       spyOn(SubnetReader, 'listSubnets').and.returnValue($q.when([]));
       spyOn(AccountService, 'getPreferredZonesByAccount').and.returnValue($q.when([]));
-      spyOn(keyPairsReader, 'listKeyPairs').and.returnValue($q.when([]));
+      spyOn(KeyPairsReader, 'listKeyPairs').and.returnValue($q.when([]));
       spyOn(awsInstanceTypeService, 'getAllTypesByRegion').and.returnValue($q.when([]));
       const refreshCacheSpy = spyOn(cacheInitializer, 'refreshCache').and.returnValue($q.when(null));
 
@@ -130,7 +127,7 @@ describe('Service: awsServerGroupConfiguration', function() {
       spyOn(loadBalancerReader, 'listLoadBalancers').and.returnValue($q.when(this.allLoadBalancers));
       spyOn(SubnetReader, 'listSubnets').and.returnValue($q.when([]));
       spyOn(AccountService, 'getPreferredZonesByAccount').and.returnValue($q.when([]));
-      spyOn(keyPairsReader, 'listKeyPairs').and.returnValue($q.when([]));
+      spyOn(KeyPairsReader, 'listKeyPairs').and.returnValue($q.when([]));
       spyOn(awsInstanceTypeService, 'getAllTypesByRegion').and.returnValue($q.when([]));
       const refreshCacheSpy = spyOn(cacheInitializer, 'refreshCache').and.returnValue($q.when(null));
 
@@ -164,7 +161,7 @@ describe('Service: awsServerGroupConfiguration', function() {
       spyOn(loadBalancerReader, 'listLoadBalancers').and.returnValue($q.when([]));
       spyOn(SubnetReader, 'listSubnets').and.returnValue($q.when([]));
       spyOn(AccountService, 'getPreferredZonesByAccount').and.returnValue($q.when([]));
-      spyOn(keyPairsReader, 'listKeyPairs').and.returnValue($q.when([]));
+      spyOn(KeyPairsReader, 'listKeyPairs').and.returnValue($q.when([]));
       const getAllTypesByRegionSpy = spyOn(awsInstanceTypeService, 'getAllTypesByRegion').and.returnValue(
         $q.when({
           'us-east-1': [{ name: 'm4.tiny' }],
